chore: remove the old RaptorQ-based encoding
