+++ Cargo.toml
@@ -86,7 +86,6 @@ proc-macro2 = "1.0.95"
 prometheus = "0.13.4"
 quote = "1.0"
 rand = "0.8.5"
-raptorq = "2.0.0"
 rayon = "1.10.0"
 rcgen = "0.13.2"
 reed-solomon-simd = "3.0.1"
