+++ crates/walrus-core/benches/blob_encoding.rs
@@ -6,10 +6,8 @@
 use core::{num::NonZeroU16, time::Duration};
 
 use criterion::{AxisScale, BatchSize, BenchmarkId, Criterion, PlotConfiguration};
-use walrus_core::encoding::{EncodingConfigTrait as _, Primary, RaptorQEncodingConfig};
+use walrus_core::encoding::{EncodingConfigTrait as _, Primary, ReedSolomonEncodingConfig};
 use walrus_test_utils::{random_data, random_subset};
-
-// TODO (WAL-610): Support both encoding types.
 
 const N_SHARDS: u16 = 1000;
 
@@ -24,8 +22,8 @@ const BLOB_SIZES: [(u64, &str); 6] = [
     (1 << 30, "1GiB"),
 ];
 
-fn encoding_config() -> RaptorQEncodingConfig {
-    RaptorQEncodingConfig::new(NonZeroU16::new(N_SHARDS).unwrap())
+fn encoding_config() -> ReedSolomonEncodingConfig {
+    ReedSolomonEncodingConfig::new(NonZeroU16::new(N_SHARDS).unwrap())
 }
 
 fn blob_encoding(c: &mut Criterion) {
