+++ crates/walrus-core/src/encoding/mapping.rs
@@ -140,11 +140,11 @@ mod tests {
     use walrus_test_utils::param_test;
 
     use super::*;
-    use crate::{encoding::RaptorQEncodingConfig, test_utils};
+    use crate::{encoding::ReedSolomonEncodingConfig, test_utils};
 
     // Fixture
     fn sliver_pairs(num: u16) -> Vec<SliverPair> {
-        let encoding_config = RaptorQEncodingConfig::new_for_test(1, 1, num);
+        let encoding_config = ReedSolomonEncodingConfig::new_for_test(1, 1, num);
         (0..num)
             .map(|n| {
                 SliverPair::new_empty(
