+++ crates/walrus-core/src/encoding/quilt_encoding.rs
@@ -21,13 +21,9 @@ use tracing::{Level, Span};
 
 use super::{EncodingConfigEnum, Primary, Secondary, SliverData, SliverPair};
 use crate::{
+    EncodingType,
     SliverIndex,
-    encoding::{
-        MAX_SYMBOL_SIZE,
-        QuiltError,
-        blob_encoding::BlobEncoder,
-        config::EncodingConfigTrait as _,
-    },
+    encoding::{QuiltError, blob_encoding::BlobEncoder, config::EncodingConfigTrait as _},
     metadata::{QuiltIndex, QuiltIndexV1, QuiltMetadata, QuiltMetadataV1, QuiltPatchV1},
 };
 
@@ -572,13 +568,12 @@ impl QuiltEncoderApi<QuiltVersionV1> for QuiltEncoderV1<'_> {
             .chain(blob_pairs.iter().map(|bwd| bwd.blob.len()))
             .collect();
 
-        let required_alignment = self.config.encoding_type().required_alignment() as usize;
         let symbol_size = utils::compute_symbol_size(
             &all_sizes,
             n_columns,
             n_rows,
             MAX_NUM_COLUMNS_FOR_QUILT_INDEX,
-            required_alignment,
+            self.config.encoding_type(),
         )?;
 
         let row_size = symbol_size * n_columns;
@@ -868,7 +863,7 @@ mod utils {
         n_columns: usize,
         n_rows: usize,
         max_num_columns_for_quilt_index: usize,
-        required_alignment: usize,
+        encoding_type: EncodingType,
     ) -> Result<usize, QuiltError> {
         if blobs_sizes.len() > n_columns {
             // The first column is not user data.
@@ -909,11 +904,13 @@ mod utils {
             }
         }
 
-        let symbol_size = min_val.next_multiple_of(required_alignment);
-        if symbol_size > MAX_SYMBOL_SIZE as usize {
+        let symbol_size = min_val.next_multiple_of(encoding_type.required_alignment().into());
+        let max_symbol_size = usize::from(encoding_type.max_symbol_size());
+        if symbol_size > max_symbol_size {
             return Err(QuiltError::QuiltOversize(format!(
-                "the resulting symbol size {} is too large, remove some blobs",
-                symbol_size
+                "the resulting symbol size {} is larger than the maximum symbol size {}; \
+                remove some blobs",
+                symbol_size, max_symbol_size
             )));
         }
 
@@ -1073,10 +1070,7 @@ mod tests {
     use walrus_test_utils::{param_test, random_data};
 
     use super::*;
-    use crate::{
-        encoding::{RaptorQEncodingConfig, ReedSolomonEncodingConfig},
-        metadata::BlobMetadataApi as _,
-    };
+    use crate::{encoding::ReedSolomonEncodingConfig, metadata::BlobMetadataApi as _};
 
     /// Get the minimum required columns.
     fn min_required_columns(blobs: &[usize], length: usize) -> usize {
@@ -1092,42 +1086,39 @@ mod tests {
 
     param_test! {
         test_quilt_find_min_length: [
-            case_1: (&[2, 1, 2, 1], 3, 3, 1, Err(QuiltError::TooManyBlobs(3, 2))),
-            case_2: (&[1000, 1, 1], 4, 7, 2, Ok(144)),
+            case_1: (&[2, 1, 2, 1], 3, 3, Err(QuiltError::TooManyBlobs(3, 2))),
+            case_2: (&[1000, 1, 1], 4, 7, Ok(144)),
             case_3: (
                 &[],
                 3,
                 1,
-                1,
                 Err(QuiltError::EmptyInput("blobs".to_string())),
             ),
-            case_4: (&[1], 3, 2, 1, Ok(3)),
-            case_5: (&[115, 80, 4], 17, 9, 1, Ok(13)),
-            case_6: (&[20, 20, 20], 3, 5, 1, Ok(4)),
-            case_7: (&[5, 5, 5], 5, 1, 2, Ok(6)),
-            case_8: (&[25, 35, 45], 200, 1, 2, Ok(26)),
-            case_9: (&[10, 0, 0, 0], 17, 9, 1, Ok(2)),
-            case_10: (&[10, 0, 0, 0], 17, 9, 2, Ok(2)),
+            case_4: (&[1], 3, 2, Ok(4)),
+            case_5: (&[115, 80, 4], 17, 9, Ok(14)),
+            case_6: (&[20, 20, 20], 3, 5, Ok(4)),
+            case_7: (&[5, 5, 5], 5, 1, Ok(6)),
+            case_8: (&[25, 35, 45], 200, 1, Ok(26)),
+            case_9: (&[10, 0, 0, 0], 17, 9, Ok(2)),
         ]
     }
     fn test_quilt_find_min_length(
         blobs: &[usize],
         n_columns: usize,
         n_rows: usize,
-        required_alignment: usize,
-        expected: Result<usize, QuiltError>,
+        expected_symbol_size_result: Result<usize, QuiltError>,
     ) {
         // Initialize tracing subscriber for this test
         let _guard = tracing_subscriber::fmt().try_init();
-        let res = utils::compute_symbol_size(
+        let symbol_size_result = utils::compute_symbol_size(
             blobs,
             n_columns,
             n_rows,
             MAX_NUM_COLUMNS_FOR_QUILT_INDEX,
-            required_alignment,
+            EncodingType::RS2,
         );
-        assert_eq!(res, expected);
-        if let Ok(min_size) = res {
+        assert_eq!(symbol_size_result, expected_symbol_size_result);
+        if let Ok(min_size) = symbol_size_result {
             assert!(min_required_columns(blobs, min_size * n_rows) <= n_columns);
         }
     }
@@ -1234,14 +1225,9 @@ mod tests {
         blobs_with_identifiers: &[BlobWithIdentifier<'_>],
         n_shards: u16,
     ) {
-        let raptorq_config = RaptorQEncodingConfig::new(NonZeroU16::try_from(n_shards).unwrap());
         let reed_solomon_config =
             ReedSolomonEncodingConfig::new(NonZeroU16::try_from(n_shards).unwrap());
 
-        construct_quilt(
-            blobs_with_identifiers,
-            EncodingConfigEnum::RaptorQ(&raptorq_config),
-        );
         construct_quilt(
             blobs_with_identifiers,
             EncodingConfigEnum::ReedSolomon(&reed_solomon_config),
@@ -1304,14 +1290,9 @@ mod tests {
     ) {
         let blobs_with_identifiers = generate_random_blobs(num_blobs, max_blob_size, min_blob_size);
 
-        let raptorq_config = RaptorQEncodingConfig::new(NonZeroU16::try_from(n_shards).unwrap());
         let reed_solomon_config =
             ReedSolomonEncodingConfig::new(NonZeroU16::try_from(n_shards).unwrap());
 
-        encode_decode_quilt(
-            &blobs_with_identifiers,
-            EncodingConfigEnum::RaptorQ(&raptorq_config),
-        );
         encode_decode_quilt(
             &blobs_with_identifiers,
             EncodingConfigEnum::ReedSolomon(&reed_solomon_config),
