+++ crates/walrus-service/daemon_openapi.yaml
@@ -368,7 +368,6 @@ components:
       type: string
       description: Supported Walrus encoding types.
       enum:
-      - RedStuffRaptorQ
       - RS2
     Epoch:
       type: integer
